Move BYOB reader to separate file
